mcp/servers/bilibili-mcp-server/requirements.txt
mcp>=1.0.0
python-dotenv>=1.0.1
